clase4

+++ src/main/java/ec/edu/itsqmet/manejo/escuela/model/Curso.java
@@ -15,7 +15,7 @@ public class Curso implements Serializable {
     private static final long serialVersionUID = 1L;
     @Id 
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long idCurso;
+    private int idCurso;
     private String cursoNombre;
     private int cursoCapacidad;
